+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 
 ### Changed
 
-- BaseButton: rework focused style.
+- BaseButton: rework style on hover and focus.
 
 - BaseHeading: type prop now accepts "small" when you need to display smaller content.
 
 - BaseProgressBar: ability to display an additional description when the user hovers.
+
+- BaseSocialLogin: rework style on hover and focus.
 
 ## 0.1.0 - 2018-11-05